requirements.txt
alembic==1.12.0
blinker==1.6.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
cryptography==41.0.3
dnspython==2.4.2
email-validator==2.0.0.post2
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
greenlet==2.0.2
gunicorn==21.2.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
packaging==23.1
psycopg2-binary==2.9.7
pycparser==2.21
python-dotenv==1.0.0
requests==2.31.0
SQLAlchemy==2.0.20
sqlalchemy-cockroachdb==2.0.1
typing_extensions==4.7.1
urllib3==2.0.4
Werkzeug==2.3.7
WTForms==3.0.1
